feat:[CI-15370]: Moved mock crane code to separate file and requested changes.
